# api-specs/openapi-spec-20251021-190000.yaml
openapi: 3.1.0
info:
  title: Decentralized Autonomous Forum API
  version: 1.0.0
  description: |
    RESTful API for the Decentralized Autonomous Forum platform. Features:
    - Point economy with crypto rewards
    - User authentication (email + OAuth2)
    - Post/comment creation with IPFS storage
    - Reputation-based moderation
    - BNB Chain integration

    **Base URL**: `https://api.forum.example.com/v1`

    **Authentication**: Session-based (HttpOnly cookies) + JWT for API clients

    **Rate Limiting**:
    - Anonymous: 100 req/hour
    - Authenticated: 1000 req/hour
    - Authenticated (Level 3+): 5000 req/hour

  contact:
    name: API Support
    email: support@forum.example.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.forum.example.com/v1
    description: Production server
  - url: https://staging-api.forum.example.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Local development server
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session cookie (HttpOnly, Secure, SameSite=Lax)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API clients

# ============================================================================
# REUSABLE SCHEMAS
# ============================================================================
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 789
        user_id:
          type: integer
          format: int64
          example: 12345
        author:
          $ref: '#/components/schemas/UserSummary'
        title:
          type: string
          minLength: 10
          maxLength: 300
          example: "How to Stake BNB on PancakeSwap: Complete Guide"
        body:
          type: string
          minLength: 10
          maxLength: 10000
          example: "In this guide, I'll walk you through..."
        body_html:
          type: string
          example: "<p>In this guide, I'll walk you through...</p>"
        slug:
          type: string
          example: "how-to-stake-bnb-on-pancakeswap-complete-guide"
        channel:
          $ref: '#/components/schemas/ChannelSummary'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        like_count:
          type: integer
          minimum: 0
          example: 234
        comment_count:
          type: integer
          minimum: 0
          example: 56
        view_count:
          type: integer
          minimum: 0
          example: 1432
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        is_pinned:
          type: boolean
          example: false
        is_locked:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-10-21T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-10-21T14:35:00Z"
        edited_at:
          type: string
          format: date-time
          nullable: true
          example: null
      required:
        - id
        - user_id
        - author
        - title
        - body
        - slug
        - like_count
        - comment_count
        - created_at
    PostSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        slug:
          type: string
        author:
          $ref: '#/components/schemas/UserSummary'
        like_count:
          type: integer
        comment_count:
          type: integer
        created_at:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 10
          maxLength: 300
        body:
          type: string
          minLength: 10
          maxLength: 10000
        channel_id:
          type: integer
          format: int64
          nullable: true
        tags:
          type: array
          maxItems: 5
          items:
            type: string
            maxLength: 50
      required:
        - title
        - body
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 10
          maxLength: 300
        body:
          type: string
          minLength: 10
          maxLength: 10000
        tags:
          type: array
          maxItems: 5
          items:
            type: string

    # ========== Comment Schemas ==========
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        post_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        author:
          $ref: '#/components/schemas/UserSummary'
        parent_id:
          type: integer
          format: int64
          nullable: true
        body:
          type: string
          minLength: 1
          maxLength: 2000
        body_html:
          type: string
        like_count:
          type: integer
          minimum: 0
        depth:
          type: integer
          minimum: 0
          maximum: 2
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        created_at:
          type: string
          format: date-time
        edited_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - post_id
        - user_id
        - author
        - body
        - like_count
        - depth
        - created_at
...
    CreateCommentRequest:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 2000
        parent_id:
          type: integer
          format: int64
          nullable: true
      required:
        - body

    # ========== Other Schemas ==========
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        level:
          type: string
        is_verified:
          type: boolean
    ChannelSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        color:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        post_count:
          type: integer
    Media:
      type: object
      properties:
        id:
          type: integer
        file_name:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        ipfs_hash:
          type: string
        ipfs_url:
          type: string
          format: uri
        width:
          type: integer
        height:
          type: integer
        alt_text:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "ValidationError"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
      required:
        - error
        - message

    # ========== Pagination Schemas ==========
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 250
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 13
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
    PostListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PostSummary'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

# ============================================================================
# API ENDPOINTS
# ============================================================================
paths:
...
  /posts:
    get:
      tags:
        - Posts
      summary: List posts
      description: |
        Get paginated list of posts with filters
        - Sort by: created_at (default), trending, top, new
        - Filter by: channel, tag, author
      operationId: listPosts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, trending, top, new]
            default: created_at
        - name: channel_id
          in: query
          schema:
            type: integer
        - name: tag
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'

    post:
      tags:
        - Posts
      summary: Create new post
      description: |
        Create a new post (costs 5 points)
        - Title: 10-300 characters
        - Body: 10-10,000 characters (markdown supported)
        - Tags: max 5 tags
      operationId: createPost
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
            example:
              title: "How to Stake BNB on PancakeSwap: Complete Guide"
              body: "In this comprehensive guide, I'll walk you through the step-by-step process..."
              channel_id: 2
              tags: ["bnb", "defi", "tutorial", "pancakeswap"]
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Insufficient points or rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{post_id}:
    get:
      tags:
        - Posts
      summary: Get post by ID
      description: Get full post with comments and author details
      operationId: getPostById
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

    patch:
      tags:
        - Posts
      summary: Update post
      description: Update post title and/or body (only within 15 minutes of creation)
      operationId: updatePost
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '403':
          description: Cannot edit post (not owner or edit window expired)
        '404':
          description: Post not found

    delete:
      tags:
        - Posts
      summary: Delete post
      description: Delete own post (points not refunded)
      operationId: deletePost
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Post deleted
        '403':
          description: Cannot delete post (not owner)
        '404':
          description: Post not found

  # ========== Comment Endpoints ==========
  /posts/{post_id}/comments:
    get:
      tags:
        - Comments
      summary: Get post comments
      description: Get all comments for a post (nested structure)
      operationId: getPostComments
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'

    post:
      tags:
        - Comments
      summary: Create comment
      description: Add a comment to a post (costs 2 points)
      operationId: createComment
      security:
        - SessionAuth: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '403':
          description: Insufficient points or rate limit exceeded
...
  /search:
    get:
      tags:
        - Search
      summary: Full-text search
      description: Search posts and users using PostgreSQL full-text search
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query
        - name: type
          in: query
          schema:
            type: string
            enum: [posts, users, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostSummary'
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

# ============================================================================
# END OF OPENAPI SPECIFICATION
# ============================================================================